backend/composite_services/transfer_tix/requirements.txt
Flask==3.0.0
Flask-PyMongo==2.3.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
certifi==2024.2.2 